Refactor into packages done
