--- BoardTest.java
@@ -1,42 +1,0 @@
-
-
-import static org.junit.Assert.*;
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Test;
-
-/**
- * The test class BoardTest.
- *
- * @author  (your name)
- * @version (a version number or a date)
- */
-public class BoardTest
-{
-    /**
-     * Default constructor for test class BoardTest
-     */
-    public BoardTest()
-    {
-    }
-
-    /**
-     * Sets up the test fixture.
-     *
-     * Called before every test case method.
-     */
-    @Before
-    public void setUp()
-    {
-    }
-
-    /**
-     * Tears down the test fixture.
-     *
-     * Called after every test case method.
-     */
-    @After
-    public void tearDown()
-    {
-    }
-}
