+++ mission/MissionContest.java
@@ -1,3 +1,5 @@
+package mission;
+
 import java.util.ArrayList;
