+++ mission/MissionContestTest.java
@@ -1,3 +1,5 @@
+package mission;
+
 import static org.junit.Assert.*;
 import org.junit.After;
 import org.junit.Before;
