+++ shapes/Canvas.java
@@ -1,3 +1,5 @@
+package shapes;
+
 import javax.swing.*;
 import java.awt.*;
 import java.util.List;
