+++ shapes/Triangle.java
@@ -1,92 +1,83 @@
+package shapes; 
+
 import java.awt.*;
-import java.awt.geom.*;
 
 /**
- * A circle that can be manipulated and that draws itself on a canvas.
+ * A triangle that can be manipulated and that draws itself on a canvas.
  * 
  * @author  Michael Kolling and David J. Barnes
- * @version 1.0.  (15 July 2000) 
+ * @version 1.0  (15 July 2000)
  */
 
-public class Circle{
-
-    public static final double PI=3.1416;
+public class Triangle{
     
-    private int diameter;
+    public static int VERTICES=3;
+    
+    private int height;
+    private int width;
     private int xPosition;
     private int yPosition;
     private String color;
     private boolean isVisible;
-    
 
-    public Circle(){
-        diameter = 30;
-        xPosition = 20;
+    /**
+     * Create a new triangle at default position with default color.
+     */
+    public Triangle(){
+        height = 30;
+        width = 40;
+        xPosition = 140;
         yPosition = 15;
-        color = "blue";
+        color = "green";
         isVisible = false;
     }
 
-
-       
+    /**
+     * Make this triangle visible. If it was already visible, do nothing.
+     */
     public void makeVisible(){
         isVisible = true;
         draw();
     }
     
-
+    /**
+     * Make this triangle invisible. If it was already invisible, do nothing.
+     */
     public void makeInvisible(){
         erase();
         isVisible = false;
     }
-
-    private void draw(){
-        if(isVisible) {
-            Canvas canvas = Canvas.getCanvas();
-            canvas.draw(this, color, 
-                new Ellipse2D.Double(xPosition, yPosition, 
-                diameter, diameter));
-            canvas.wait(10);
-        }
-    }
-
-    private void erase(){
-        if(isVisible) {
-            Canvas canvas = Canvas.getCanvas();
-            canvas.erase(this);
-        }
-    }
     
     /**
-     * Move the circle a few pixels to the right.
+     * Move the triangle a few pixels to the right.
      */
     public void moveRight(){
         moveHorizontal(20);
     }
 
     /**
-     * Move the circle a few pixels to the left.
+     * Move the triangle a few pixels to the left.
      */
     public void moveLeft(){
         moveHorizontal(-20);
     }
 
     /**
-     * Move the circle a few pixels up.
+     * Move the triangle a few pixels up.
      */
     public void moveUp(){
         moveVertical(-20);
     }
 
     /**
-     * Move the circle a few pixels down.
+     * Move the triangle a few pixels down.
      */
     public void moveDown(){
         moveVertical(20);
     }
 
     /**
-     * Move the circle horizontally.
+     * Move the triangle horizontally.
      * @param distance the desired distance in pixels
      */
     public void moveHorizontal(int distance){
@@ -96,7 +87,7 @@ public class Circle{
     }
 
     /**
-     * Move the circle vertically.
+     * Move the triangle vertically.
      * @param distance the desired distance in pixels
      */
     public void moveVertical(int distance){
@@ -106,7 +97,7 @@ public class Circle{
     }
 
     /**
-     * Slowly move the circle horizontally.
+     * Slowly move the triangle horizontally.
      * @param distance the desired distance in pixels
      */
     public void slowMoveHorizontal(int distance){
@@ -126,7 +117,7 @@ public class Circle{
     }
 
     /**
-     * Slowly move the circle vertically
+     * Slowly move the triangle vertically.
      * @param distance the desired distance in pixels
      */
     public void slowMoveVertical(int distance){
@@ -135,7 +126,7 @@ public class Circle{
         if(distance < 0) {
             delta = -1;
             distance = -distance;
-        }else {
+        } else {
             delta = 1;
         }
 
@@ -146,15 +137,17 @@ public class Circle{
     }
 
     /**
-     * Change the size.
-     * @param newDiameter the new size (in pixels). Size must be >=0.
+     * Change the size to the new size
+     * @param newHeight the new height in pixels. newHeight must be >=0.
+     * @param newWidht the new width in pixels. newWidht must be >=0.
      */
-    public void changeSize(int newDiameter){
+    public void changeSize(int newHeight, int newWidth) {
         erase();
-        diameter = newDiameter;
+        height = newHeight;
+        width = newWidth;
         draw();
     }
-
+    
     /**
      * Change the color. 
      * @param color the new color. Valid colors are "red", "yellow", "blue", "green",
@@ -165,6 +158,26 @@ public class Circle{
         draw();
     }
 
+    /*
+     * Draw the triangle with current specifications on screen.
+     */
+    private void draw(){
+        if(isVisible) {
+            Canvas canvas = Canvas.getCanvas();
+            int[] xpoints = { xPosition, xPosition + (width/2), xPosition - (width/2) };
+            int[] ypoints = { yPosition, yPosition + height, yPosition + height };
+            canvas.draw(this, color, new Polygon(xpoints, ypoints, 3));
+            canvas.wait(10);
+        }
+    }
 
-
+    /*
+     * Erase the triangle on screen.
+     */
+    private void erase(){
+        if(isVisible) {
+            Canvas canvas = Canvas.getCanvas();
+            canvas.erase(this);
+        }
+    }
 }
