+++ shapes/package.bluej
@@ -1,0 +1,57 @@
+#BlueJ package file
+dependency1.from=Triangle
+dependency1.to=Canvas
+dependency1.type=UsesDependency
+dependency2.from=Circle
+dependency2.to=Canvas
+dependency2.type=UsesDependency
+dependency3.from=Rectangle
+dependency3.to=Canvas
+dependency3.type=UsesDependency
+objectbench.height=93
+objectbench.width=762
+package.divider.horizontal=0.6
+package.divider.vertical=0.8007968127490039
+package.editor.height=395
+package.editor.width=646
+package.editor.x=1612
+package.editor.y=184
+package.frame.height=610
+package.frame.width=804
+package.numDependencies=3
+package.numTargets=4
+package.showExtends=true
+package.showUses=true
+readme.height=60
+readme.name=@README
+readme.width=49
+readme.x=10
+readme.y=10
+target1.height=70
+target1.name=Circle
+target1.showInterface=false
+target1.type=ClassTarget
+target1.width=120
+target1.x=280
+target1.y=60
+target2.height=70
+target2.name=Canvas
+target2.showInterface=false
+target2.type=ClassTarget
+target2.width=120
+target2.x=20
+target2.y=140
+target3.height=70
+target3.name=Rectangle
+target3.showInterface=false
+target3.type=ClassTarget
+target3.width=120
+target3.x=280
+target3.y=250
+target4.height=70
+target4.name=Triangle
+target4.showInterface=false
+target4.type=ClassTarget
+target4.width=120
+target4.x=140
+target4.y=250
